refactor: printSum인자를 int타입에서 long타입으로 변경

+++ src/main/java/calculator/OutputHandler.java
@@ -6,7 +6,7 @@ public class OutputHandler {
         System.out.println("덧셈할 문자열을 입력해 주세요.");
     }
 
-    public void printSum(int sum) {
+    public void printSum(long sum) {
         System.out.println("결과 : " + sum);
     }
